refactor: auth exception & success 파일 적용
